requirements.txt
google-cloud-speech>=2.21.0
google-auth>=2.25.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
pyaudio>=0.2.13
pynput>=1.7.6
elevenlabs>=0.2.0